0.01.81
Zadania cd

+++ CosmicSpaceCommunication/Game/Quest/PilotTaskQuest.cs
@@ -7,7 +7,7 @@ namespace CosmicSpaceCommunication.Game.Quest
     public class PilotTaskQuest
     {
         public ulong Id { get; set; }
-        public ulong Progress { get; set; }
+        public decimal Progress { get; set; }
         public bool IsDone { get; set; }
 
         public Quest Quest { get; set; }
@@ -19,7 +19,7 @@ namespace CosmicSpaceCommunication.Game.Quest
             return new PilotTaskQuest()
             {
                 Id = ConvertRow.Row<ulong>(row["pilottaskquestid"]),
-                Progress = ConvertRow.Row<ulong>(row["progress"]),
+                Progress = ConvertRow.Row<decimal>(row["progress"]),
                 IsDone = ConvertRow.Row<int>(row["isdone"]) == 1,
             };
         }
